organiza o interfaces e corrige tipagens
